fix flux measurement

+++ heuslertools/tools/data_handling.py
@@ -14,13 +14,13 @@ def search_data_start(file, identifier, encoding):
             return i
         i += 1
 
-def load_data(file, identifier, delimiter, start_row=0, end_row=None, names=True, encoding=None):
+def load_data(file, identifier, delimiter, start_row=0, end_row=None, names=True, encoding=None, dtype=float):
     """
     Loads a measurement and returns the data
     moment data.
     """
     data = {}
-    data = np.genfromtxt(fname=file, delimiter=delimiter, skip_header=search_data_start(file, identifier, encoding), names=names, encoding=encoding)
+    data = np.genfromtxt(fname=file, delimiter=delimiter, skip_header=search_data_start(file, identifier, encoding), names=names, encoding=encoding, dtype=dtype)
     first_line = [str(i) for i in data[0]]
     if all(x == 'nan' for x in first_line):
         data = data[1:]
